Get entire region at once.
We were getting the whole image at full resolution a tile at time.  This
just gets the image all at once, which is fewer calls to the server at
the expense of possible parallelization of the tiles.  Eventually we
will switch to a slippy-map viewer (geojs) and let it choose how to tile
so that we could deal with high-resolution images and tiling XY panels.