Always use the histogram for image style.
We automatically adjust images so that the min and max of the histogram
informs the range of the style, but this only works if we have already
loaded the histogram.

This could be improved further -- we will fetch one image per histogram,
even if the use has changed something so the histogram is no longer
needed.