Hide the start upload button
This is being done as a proof-of-concept. Setting the start button text
to the empty string does not prevent the button from being clicked
(however, see https://github.com/girder/girder_web_components/pull/234).
But for now we're doing it this way to allow us to see what this could
look like.