Track configuration creation locally
The store action that creates the new configuration does *not* update
the store itself by committing a mutation, etc. The quick fix is to hold
on to a singleton instance of the configuration locally for use in
updating the template immediately, and relying on the correct values
being in the store on the next load of this view.

This is clearly not an optimal way to handle this; the solution is to
use a more solid approach to application state tracking.