Reduce the number of images fetched.
We were fetching a full range image for the histogram preview.  If this
is fetched before the histogram, it used query parameters of min: min,
max: max.  If fetched after the histogram, it specified numerical
values.  This always prefers min/max in the query for full range.

Also, reduce the number of palette steps -- we are using a linear
transform, so additional steps are of no benefit (and could introduce
mild artifacts for colors that are do not use the full possible range).