Use the quad_info endpoint for tile_frames.
Now, precaching the tile frames can be done by calling the
tile_frames/quad_info endpoint with cache=schedule.  To match what is
done in the javascript side, this endpoint needs to be called per
channel with frameBase=<channel>, frameStride=<number of channels>,
frameGroup=<IndexZ || 1>, frameGroupStride=<IndexZ / IndexC>.  This is
in addition to the defaults used in this project defined in
src/store/index.ts around line 745.

To make this more convenient, we could add (to large_image) a special
value for frameBase which would schedule all tile_frames based on this
heuristic.