Increase the tile queue tracking parameter.
The default is large enough for a single layer, but we have lots of
layers.  Making it bigger uses more browser memory but prevents trying
to discard tiles needlessly.  Ideally, we'd modify geojs so when layers
share a tile queue the tracking parameter would be a per-lyater value
not a fixed value.