Improve the test environment.
Specifically, the Girder plugin setup.py file was missing some
requirements.  The tox file should specify using find-links to fetch
large_image wheels.  The Dockerfile now more conveniently starts in the
directory where the plugin is located.  Lastly, there is a test_env.sh
script that can be run in the docker to allow girder tests to properly
find mongo (unfortunately, girder tests expect mongo to be on localhost,
which isn't the case in a docker-compose environment).