Change when we schedule caching tile_frames and maxmerge
We were scheduling these when a file was uploaded, but then converting
that item into a different record via the multi-source file.  This
largely invalidated the usefulness of the tile_frames and maxmerge,
especially if the multi-source file was not a direct pass through.

This moves the scheduling of these to after the multi-source file
becomes a large image or is transcoded into a tiff.