Fix networking mode for worker.
The worker needs to run in the same network as the rest of the docker
containers OR girder must be publicly visible to the worker and to
girder_worker both.  By not specifying the network mode of the worker,
girder_worker assigns it the same network as girder_worker's container.
By specifying http://girder:8080/api/v1 as the alternative worker URL,
this is reachable from within that network.