Set `crossOrigin` to `"use-credentials"` for all images
This change requires that you set the `core.cors.allow_origin` setting in Girder.
It was a wildcard `*` but this doesn't work with `use-credentials`, so it is now `http://localhost:*`.
This setting can be changed in the Advanced Settings of the System configuration page in Girder.
You can also use a list of allowed origins which can be any unix shell pattern like `**`.
In production, you could use a reverse proxy, so that CORS are not a problem anymore.